fix(android): map numeric theme scale to Compose units

+++ apps/android/app/src/main/kotlin/dev/crosslab/android/features/appearance/ComposeTheme.kt
@@ -37,8 +37,8 @@ fun OklchColor.toComposeColor(): Color {
 fun TextScale.toTextStyle(): TextStyle =
     TextStyle(
         fontFamily = FontFamily.SansSerif,
-        fontSize = size.sp,
-        lineHeight = lineHeight.sp,
+        fontSize = size.toFloat().sp,
+        lineHeight = lineHeight.toFloat().sp,
         fontWeight = FontWeight(weight),
     )
